+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Change return type of `getIterator` method to generic `iterable`
 
 ## 3.9.0 - 2018-07-24
 - Add `containsBy` method to `ICollection`